deleted unused files and moved table form to popup folder
